backend/requirements.txt
fastapi
python-dotenv
uvicorn
kokoro
soundfile
pydantic_ai
pydantic
httpx
pymongo
faster-whisper
passlib
google-genai